+++ backend/src/main/java/com/golden_dobakhe/HakPle/domain/user/user/controller/ApiV1RegistController.java
@@ -1,11 +1,9 @@
 package com.golden_dobakhe.HakPle.domain.user.user.controller;
 
 import com.golden_dobakhe.HakPle.domain.user.user.WithdrawResult;
-import com.golden_dobakhe.HakPle.domain.user.user.dto.ChangePasswordRequest;
 import com.golden_dobakhe.HakPle.domain.user.user.dto.DeleteUserRequestDto;
-import com.golden_dobakhe.HakPle.domain.user.user.dto.ResetPasswordRequest;
-import com.golden_dobakhe.HakPle.domain.user.user.dto.UserDTO;
-import com.golden_dobakhe.HakPle.domain.user.user.service.UserService;
+import com.golden_dobakhe.HakPle.domain.user.user.dto.UserRegistRequestDTO;
+import com.golden_dobakhe.HakPle.domain.user.user.service.UserRegistService;
 import com.golden_dobakhe.HakPle.security.CustomUserDetails;
 import com.golden_dobakhe.HakPle.security.jwt.JwtTokenizer;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,29 +12,35 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-import java.util.stream.Collectors;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/users")
-@Tag(name = "회원가입 API", description = "사용자 관련 API")
-public class ApiV1UserController {
-    private final UserService userService;
+@Tag(name = "회원가입 API", description = "회원가입, 탈퇴, 로그아웃 API")
+public class ApiV1RegistController {
+    private final UserRegistService userRegistService;
     private final JwtTokenizer jwtTokenizer;
     private final RedisTemplate<String, String> redisTemplate;
 
+    // 회원 가입
     @Operation(summary = "회원가입", description = "유저 정보를 입력하여 회원가입을 진행합니다.")
     @PostMapping("/userreg")
     public ResponseEntity<String> register(
             @Parameter(description = "회원가입 정보", required = true)
-            @Valid @RequestBody UserDTO userDTO,
+            @Valid @RequestBody UserRegistRequestDTO userRegistRequestDTO,
             BindingResult bindingResult
     ) {
         if (bindingResult.hasErrors()) {
@@ -47,51 +51,22 @@ public class ApiV1UserController {
             );
         }
 
-        userService.register(userDTO);
+        userRegistService.register(userRegistRequestDTO);
         return ResponseEntity.ok("회원가입이 성공적으로 완료되었습니다.");
     }
 
-    @Operation(summary = "비밀번호 변경", description = "로그인된 사용자가 현재 비밀번호를 확인하고 새 비밀번호로 변경합니다.")
-    @PostMapping("/change-password")
-    public ResponseEntity<?> changePassword(
-            @AuthenticationPrincipal CustomUserDetails principal,
-            @RequestBody @Valid ChangePasswordRequest request
-    ) {
-        userService.changePasswordWithOldPassword(
-                principal.getUser().getId(),
-                request.getCurrentPassword(),
-                request.getNewPassword(),
-                request.getNewPasswordConfirm()
-        );
-        return ResponseEntity.ok("비밀번호가 변경되었습니다.");
-    }
-
-    @Operation(summary = "비밀번호 재설정", description = "비밀번호 찾기 이후 인증 완료된 사용자가 새 비밀번호를 설정합니다.")
-    @PostMapping("/reset-password")
-    public ResponseEntity<?> resetPassword(
-            @Parameter(description = "사용자 ID", example = "1")
-            @AuthenticationPrincipal CustomUserDetails principal,
-            @RequestBody @Valid ResetPasswordRequest request
-    ) {
-        userService.resetPassword(
-                principal.getUser().getId(),
-                request.getNewPassword(),
-                request.getNewPasswordConfirm()
-        );
-        return ResponseEntity.ok("비밀번호가 재설정되었습니다.");
-    }
-
-    @DeleteMapping("/withdraw")
+    // 탈퇴
     @Operation(summary = "회원 탈퇴", description = "비밀번호 확인 후 회원 비활성화 처리")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "회원 탈퇴 성공"),
             @ApiResponse(responseCode = "401", description = "비밀번호 불일치"),
             @ApiResponse(responseCode = "404", description = "사용자를 찾을 수 없음")
     })
+    @DeleteMapping("/withdraw")
     public ResponseEntity<String> withdrawUser(@RequestBody DeleteUserRequestDto dto,
                                                @AuthenticationPrincipal CustomUserDetails principal) {
         Long userId = principal.getUser().getId();
-        WithdrawResult result = userService.withdraw(userId, dto.getPassword());
+        WithdrawResult result = userRegistService.withdraw(userId, dto.getPassword());
 
         return switch (result) {
             case SUCCESS -> ResponseEntity.ok("회원 탈퇴 완료");
@@ -100,7 +75,7 @@ public class ApiV1UserController {
         };
     }
 
-
+    // 로그아웃
     @Operation(summary = "로그아웃", description = "Access Token을 블랙리스트에 등록하고 Refresh Token을 제거합니다.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "로그아웃 성공"),
@@ -113,13 +88,8 @@ public class ApiV1UserController {
         String token = bearerToken.replace("Bearer ", "");
         Long userId = principal.getUser().getId();
 
-        userService.logout(token, userId);
+        userRegistService.logout(token, userId);
 
         return ResponseEntity.ok("로그아웃 완료");
     }
-
-
-
-
-
 }
